==> notebooks/04_data_splits.ipynb <==
# Data splits for cross-validation

import os
import time
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from glob import glob
from tqdm import tqdm
import os

src_date = "2023-06-17"
src_dir = f"../datasets/processed/post-review/{src_date}"

import psutil
    
ram_gb = psutil.virtual_memory().total / 2**30 # total physical memory in bytes
print(f"RAM: {ram_gb:.2f} GB")

## Load data

# # set up columns : DONT for ease of eval
# time_cols = ['datetime','hour', 'dayofweek']

# road_cols = [
#     'number_of_lanes', 'speed_limit_kph', 'elevation',
# ]

# vehicle_cols = [
#     'agg_speed', 
# #     'lat', 'lon', 'vehicle_id', 'num_periods', 'barangay',
# ]

# landuse_cols = []
# buffer_sizes = [50, 100, 150, 200, 250, 500, 1000]
# for size in buffer_sizes:
#     landuse_cols.extend(
#         [
#             f'pix_residential_4x4_{size}',
#             f'pix_institutional_4x4_{size}',
#             f'pix_industrial_4x4_{size}',
#             f'pix_business_4x4_{size}',
#         ]
#     )
# usecols = time_cols + road_cols + vehicle_cols + landuse_cols

paths = glob(os.path.join(src_dir, "*.csv"))
df = pd.DataFrame()
for path in tqdm(paths):
    df = pd.concat([df, pd.read_csv(path)])

# sort by datetime
df["datetime"] = pd.to_datetime(df["datetime"])
df.sort_values(by="datetime", inplace=True)

## Split according to time

from sklearn.model_selection import TimeSeriesSplit, train_test_split

train, test = train_test_split(
    df, test_size=0.20, random_state=42, shuffle=False
)

print("Train shape: ", train.shape)
print("Test shape: ", test.shape)

(train["datetime"].min(), train['datetime'].max(),
 test["datetime"].min(), test["datetime"].max())

save_dir = f"../datasets/processed/post-review-tt-splits/{src_date}"
os.makedirs(save_dir, exist_ok=True)

train.to_csv(os.path.join(save_dir, "train.csv"), index=False)
test.to_csv(os.path.join(save_dir, "test.csv"), index=False)

tscv = TimeSeriesSplit(n_splits=3)

for i, (train_index, test_index) in enumerate(tscv.split(train)):
    fold_train = train.iloc[train_index]
    fold_test = train.iloc[test_index]
    
    print(f"Fold {i} train shape: ", fold_train.shape)
    print(f"Fold {i} test shape: ", fold_test.shape)
    
    save_dir = f"../datasets/processed/post-review-ts-splits/{src_date}"
    os.makedirs(save_dir, exist_ok=True)
    fold_train.to_csv(
        os.path.join(save_dir, f"fold-{i}-train.csv"), index=False)
    fold_test.to_csv(
        os.path.join(save_dir, f"fold-{i}-test.csv"), index=False)
    
    # use first fold as tuning set, save to tt
    save_dir = f"../datasets/processed/post-review-tt-splits/{src_date}"
    if not i:
        tune = fold_train
        tune.to_csv(os.path.join(save_dir, "tune.csv"), index=False) # 1/3 of train

tune['datetime'].sort_values()
